@@ -1,84 +1,76 @@
 {
-  "name": "project-strokes",
-  "description": "",
-  "version": "0.1.0",
+  "name": "albavulpes-api",
+  "description": "Project Strokes - Web API Server",
+  "version": "0.0.0",
+  "private": true,
   "copyright": "Copyright © 2017",
   "author": {
     "name": "cypherix93",
-    "url": "https://github.com/cypherix93/project-strokes"
+    "url": "https://github.com/cypherix93"
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/albavulpes/project-strokes-server.git"
+    "url": "https://github.com/albavulpes/api.git"
   },
-  "private": true,
   "license": "UNLICENSED",
   "dependencies": {
     "bcryptjs": "^2.3.0",
     "body-parser": "^1.14.2",
     "cookie-parser": "^1.4.1",
     "cors": "^2.7.1",
-    "emailjs": "^1.0.2",
     "express": "^4.14.0",
     "express-session": "^1.14.1",
     "http-status-codes": "^1.0.6",
-    "hydrate-mongodb": "^1.3.3",
+    "hydrate-mongodb": "^1.8.0",
     "jsonwebtoken": "^7.1.9",
     "lodash": "^4.16.3",
     "moment": "^2.11.1",
     "mongodb": "^2.2.19",
-    "node-uuid": "^1.4.7",
-    "passport": "^0.3.2",
+    "passport": "^0.4.0",
     "passport-local": "^1.0.0",
-    "q": "^1.4.1",
-    "routing-controllers": "^0.6.9",
-    "shortid": "^2.2.6",
-    "validator": "^6.2.1",
+    "routing-controllers": "^0.7.1",
+    "shortid": "^2.2.8",
+    "validator": "^8.1.0",
     "winston": "^2.3.1"
   },
   "devDependencies": {
     "@types/bcryptjs": "^2.3.30",
-    "@types/body-parser": "0.0.33",
+    "@types/body-parser": "^1.16.5",
     "@types/cookie-parser": "^1.3.30",
     "@types/cors": "^2.8.0",
     "@types/express": "^4.0.33",
-    "@types/express-session": "0.0.32",
+    "@types/express-session": "^1.15.3",
     "@types/http-status-codes": "^1.0.29",
     "@types/jsonwebtoken": "^7.1.32",
     "@types/lodash": "^4.14.36",
-    "@types/node": "^7.0.4",
+    "@types/node": "^8.0.25",
     "@types/node-uuid": "0.0.28",
     "@types/passport": "^0.3.2",
     "@types/passport-local": "^1.0.29",
-    "@types/q": "0.0.32",
-    "@types/shortid": "0.0.28",
-    "@types/validator": "^5.7.32",
+    "@types/shortid": "^0.0.29",
+    "@types/validator": "^6.2.2",
     "@types/winston": "^2.2.0",
-    "del": "^2.2.0",
-    "gulp": "^3.9.0",
-    "gulp-batch": "^1.0.5",
+    "del": "^3.0.0",
+    "gulp": "^3.9.1",
     "gulp-cached": "^1.1.1",
-    "gulp-debug": "^3.0.0",
-    "gulp-filter": "^5.0.0",
-    "gulp-json-editor": "^2.2.1",
-    "gulp-load-plugins": "^1.2.0",
-    "gulp-nodemon": "^2.1.0",
-    "gulp-plumber": "^1.0.1",
-    "gulp-rename": "^1.2.2",
-    "gulp-run": "^1.6.12",
-    "gulp-typescript": "^3.1.4",
-    "gulp-util": "^3.0.7",
-    "gulp-watch": "^4.3.5",
-    "gulp-zip": "^4.0.0",
-    "merge-stream": "^1.0.0",
-    "run-sequence": "^1.1.5",
-    "streamqueue": "^1.1.1",
-    "typescript": "^2.1.5"
+    "gulp-concat": "^2.6.1",
+    "gulp-debug": "^3.1.0",
+    "gulp-load-plugins": "^1.5.0",
+    "gulp-run": "^1.7.1",
+    "gulp-sourcemaps": "^2.6.0",
+    "gulp-typescript": "^3.1.7",
+    "gulp-watch": "^4.3.11",
+    "merge-stream": "^1.0.1",
+    "run-sequence": "^2.1.0",
+    "tslint": "^5.1.0",
+    "typescript": "^2.4.2"
   },
   "scripts": {
     "gulp": "./node_modules/.bin/gulp",
+    "clean": "npm run gulp clean",
     "build": "npm run gulp build",
-    "package": "npm run gulp package",
-    "start": "npm run gulp start"
+    "watch": "npm run gulp watch",
+    "start": "nodemon ./dist/Server.js --watch ./dist/",
+    "test": "echo \"Error: no test specified\" && exit 1"
   }
 }
