@@ -1,4 +1,4 @@
-import path = require("path");
-import {CONFIG} from "./Config";
+import path = require('path');
+import {CONFIG} from './Config';
 
-export const IMAGES_STORAGE_PATH = path.join(CONFIG.settings.fileStoragePath, "images");
+export const IMAGES_STORAGE_PATH = path.join(CONFIG.settings.fileStoragePath, 'images');
