@@ -1,9 +1,8 @@
-import _ = require("lodash");
-import {SessionFactory} from "hydrate-mongodb";
+import _ = require('lodash');
+import {SessionFactory} from 'hydrate-mongodb';
 
-export var SessionManager = {} as SessionFactory;
+export const SessionManager = {} as SessionFactory;
 
-export function initSessionManager(sessionFactory: SessionFactory)
-{
+export function initSessionManager(sessionFactory: SessionFactory) {
     _.assignIn(SessionManager, sessionFactory);
 }
