@@ -1,36 +1,32 @@
-import moment = require("moment");
-import * as jwt from "jsonwebtoken";
+import moment = require('moment');
+import * as jwt from 'jsonwebtoken';
 
-import {CONFIG} from "../config/Config";
-import {SignOptions} from "jsonwebtoken";
+import {CONFIG} from '../config/Config';
+import {SignOptions} from 'jsonwebtoken';
 
-import {User} from "../database/models/auth/User";
+import {User} from '../database/models/auth/User';
 
-export interface IPassportUser
-{
+export interface IPassportUser {
     id: string,
     email: string,
     roles: string[],
     iat?: number;
 }
 
-export class AuthHelper
-{
+export class AuthHelper {
     // Helper that logs in a user and sets them to the session
-    public static registerUserToSession(user: User, request, response, next)
-    {
+    public static registerUserToSession(user: User, request, response, next) {
         // Form the user to set to session (we dont wanna set all the user data to session)
-        var passportUser = AuthHelper.getUserForSession(user);
+        const passportUser = AuthHelper.getUserForSession(user);
 
         // Set the iat for the logged in user
         passportUser.iat = moment().unix().valueOf();
 
         // At this point, we are authenticated, so lets generate a JWT for the user
-        var jwtToken = AuthHelper.generateJWToken(passportUser);
+        const jwtToken = AuthHelper.generateJWToken(passportUser);
 
         // Set user to session
-        request.login(passportUser, function (err)
-        {
+        request.login(passportUser, function (err) {
             if (err)
                 return next(err);
 
@@ -41,8 +37,7 @@ export class AuthHelper
     }
 
     // Helper that sets a user to session
-    public static getUserForSession(user: User): IPassportUser
-    {
+    public static getUserForSession(user: User): IPassportUser {
         return {
             id: user.id,
             email: user.email,
@@ -50,24 +45,21 @@ export class AuthHelper
         };
     }
 
-    public static generateJWToken(passportUser: IPassportUser, options?: SignOptions)
-    {
+    public static generateJWToken(passportUser: IPassportUser, options?: SignOptions) {
         options = options || {};
 
         return jwt.sign(passportUser, CONFIG.jwt.secret, options);
     }
 
-    public static setAuthCookie(token, request, response)
-    {
-        var cookieConfig = CONFIG.jwt.cookie;
+    public static setAuthCookie(token, request, response) {
+        const cookieConfig = CONFIG.jwt.cookie;
 
         response.cookie(cookieConfig.name, token, cookieConfig.options);
 
         request.token = token;
     }
 
-    public static clearAuthCookie(response)
-    {
+    public static clearAuthCookie(response) {
         return response.clearCookie(CONFIG.jwt.cookie.name);
     }
 }
