@@ -1,22 +1,20 @@
-import "reflect-metadata"; // this shim is required
-import path = require("path");
+import 'reflect-metadata'; // this shim is required
+import path = require('path');
 
-import {Express} from "express";
-import {useExpressServer} from "routing-controllers";
+import {Express} from 'express';
+import {useExpressServer} from 'routing-controllers';
 
-import {CONFIG} from "../../config/Config";
-import {ErrorsConfig} from "./ErrorsConfig";
-import {LOGGER} from "../../helpers/Logger";
+import {CONFIG} from '../../config/Config';
+import {ErrorsConfig} from './ErrorsConfig';
+import {LOGGER} from '../../helpers/Logger';
 
-export class RoutingBootstrapper
-{
-    public static async init(app:Express)
-    {
-        LOGGER.info("Setting up Routes...");
+export class RoutingBootstrapper {
+    public static async init(app: Express) {
+        LOGGER.info('Setting up Routes...');
 
         // Setup routing-controllers
         useExpressServer(app, {
-            controllers: [path.join(CONFIG.rootPath, "api/controllers/**/*.js")]
+            controllers: [path.join(CONFIG.rootPath, 'api/controllers/**/*.js')]
         });
 
         // Setup error handlers for all routes
