@@ -1,14 +1,12 @@
-import * as _ from "lodash";
-import * as HttpStatus from "http-status-codes";
+import * as _ from 'lodash';
+import * as HttpStatus from 'http-status-codes';
 
-import {RequestHandler} from "express";
-import {AuthWorker} from "../workers/AuthWorker";
+import {RequestHandler} from 'express';
+import {AuthWorker} from '../workers/AuthWorker';
 
 // Use this to make routes require authorization
-export function authorize(...roles: string[]): RequestHandler
-{
-    return function (request, response, next)
-    {
+export function authorize(...roles: string[]): RequestHandler {
+    return function (request, response, next) {
         // User hasn't logged in, so send 401
         if (!request.user)
             return response.sendStatus(HttpStatus.UNAUTHORIZED);
@@ -19,7 +17,7 @@ export function authorize(...roles: string[]): RequestHandler
             return next();
 
         // Roles were requested, User has logged in, let's check roles
-        var userRoles = request.user.roles;
+        const userRoles = request.user.roles;
 
         // None of user's roles match the roles requested, send 401
         if (_.intersection(roles, userRoles).length === 0)
@@ -31,14 +29,12 @@ export function authorize(...roles: string[]): RequestHandler
 }
 
 // Use this to perform register on the register endpoint
-export function register(): RequestHandler
-{
+export function register(): RequestHandler {
     return (request, response, next) => AuthWorker.doRegister(request, response, next);
 }
 
 // Use this to perform login on the login endpoint
-export function login(): RequestHandler
-{
+export function login(): RequestHandler {
     return (request, response, next) => AuthWorker.doLogin(request, response, next);
 }
 
