@@ -1,44 +1,38 @@
-import * as moment from "moment";
-import * as jwt from "jsonwebtoken";
+import * as moment from 'moment';
+import * as jwt from 'jsonwebtoken';
 
-import {CONFIG} from "../../config/Config";
-import {AuthHelper, IPassportUser} from "../../helpers/AuthHelper";
-import {SessionManager} from "../../database/SessionManager";
-import {User} from "../../database/models/auth/User";
-import {IPayload} from "../../interfaces/IPayload";
+import {CONFIG} from '../../config/Config';
+import {AuthHelper, IPassportUser} from '../../helpers/AuthHelper';
+import {SessionManager} from '../../database/SessionManager';
+import {User} from '../../database/models/auth/User';
+import {IPayload} from '../../interfaces/IPayload';
 
 // Middleware to hook up JWT
-export function jsonWebToken()
-{
-    return function (request, response, next)
-    {
+export function jsonWebToken() {
+    return function (request, response, next) {
         // check cookies or url parameters or post parameters for token
-        var token = request.body.token || request.query.token || request.cookies[CONFIG.jwt.cookie.name];
+        let token = request.body.token || request.query.token || request.cookies[CONFIG.jwt.cookie.name];
 
         if (!token)
             return next();
 
         // Verifies secret and checks expiry
-        jwt.verify(token, CONFIG.jwt.secret, function (err, decoded: IPassportUser)
-        {
+        jwt.verify(token, CONFIG.jwt.secret, function (err, decoded: IPassportUser) {
             if (err)
                 return next(err);
 
             // Let's refresh the token if it has expired
             refreshTokenIfExpired(decoded)
-                .then(resolved =>
-                {
-                    var data = resolved.data;
+                .then(resolved => {
+                    const data = resolved.data;
 
                     // If new token was provided, update cookie
-                    if (data.token)
-                    {
+                    if (data.token) {
                         AuthHelper.setAuthCookie(data.token, request, response);
                     }
 
                     // If user was removed, clear cookie
-                    if (!data.decoded)
-                    {
+                    if (!data.decoded) {
                         AuthHelper.clearAuthCookie(response);
                     }
 
@@ -48,40 +42,36 @@ export function jsonWebToken()
                     // Carry on
                     return next();
                 })
-                .catch(function (err)
-                {
+                .catch(function (err) {
                     return next(err);
                 });
         });
-    }
+    };
 }
 
-async function refreshTokenIfExpired(decoded: IPassportUser): Promise<IPayload<{decoded: IPassportUser, token?: string}>>
-{
-    var issuedAt = decoded.iat;
-    var now = moment().unix().valueOf();
+async function refreshTokenIfExpired(decoded: IPassportUser): Promise<IPayload<{ decoded: IPassportUser, token?: string }>> {
+    const issuedAt = decoded.iat;
+    const now = moment().unix().valueOf();
 
-    var expiryDuration = CONFIG.jwt.expiryInMinutes * 60;
+    const expiryDuration = CONFIG.jwt.expiryInMinutes * 60;
 
     // If not expired just return the previous decoded token
-    if ((now - issuedAt) <= expiryDuration)
-    {
+    if ((now - issuedAt) <= expiryDuration) {
         return {
             success: true,
             data: {decoded: decoded}
         };
     }
 
-    var session = SessionManager.createSession();
+    const session = SessionManager.createSession();
 
     // Otherwise, let's recheck the database and make sure User claims the correct stuff
-    var dbUser = await session.query(User).findOne({_id: decoded.id}).asPromise();
+    let dbUser = await session.query(User).findOne({_id: decoded.id}).asPromise();
 
     session.close();
 
     // If user does not exist anymore, invalidate the session user
-    if (!dbUser)
-    {
+    if (!dbUser) {
         return {
             success: false,
             data: {decoded: null}
@@ -101,5 +91,5 @@ async function refreshTokenIfExpired(decoded: IPassportUser): Promise<IPayload<{
             decoded: decoded,
             token: AuthHelper.generateJWToken(decoded)
         }
-    }
+    };
 }
