@@ -1,47 +1,43 @@
-import {IPayload} from "../../../interfaces/IPayload";
-import {IRestWorker} from "../../../interfaces/IRestWorker";
-import {SessionManager} from "../../../database/SessionManager";
-import {Season} from "../../../database/models/comic/Season";
-import {Arc} from "../../../database/models/comic/Arc";
+import {IPayload} from '../../../interfaces/IPayload';
+import {IRestWorker} from '../../../interfaces/IRestWorker';
+import {SessionManager} from '../../../database/SessionManager';
+import {Season} from '../../../database/models/comic/Season';
+import {Arc} from '../../../database/models/comic/Arc';
 
-class Worker implements IRestWorker<Arc>
-{
-    public async create(body: Arc): Promise<IPayload<Arc>>
-    {
+class Worker implements IRestWorker<Arc> {
+
+    public async create(body: Arc): Promise<IPayload<Arc>> {
         // Validate request
-        if (!body.seasonId)
-        {
+        if (!body.seasonId) {
             return {
                 success: false,
-                message: "Season information given is invalid. Cannot create Arc."
+                message: 'Season information given is invalid. Cannot create Arc.'
             };
         }
 
-        var session = SessionManager.createSession();
+        const session = SessionManager.createSession();
 
         // Get the proper parent
-        var season = await session.query(Season).findOne({_id: body.seasonId}).asPromise();
+        let season = await session.query(Season).findOne({_id: body.seasonId}).asPromise();
 
-        if (!season)
-        {
+        if (!season) {
             return {
                 success: false,
-                message: "Requested Season does not exist. Cannot create Arc."
+                message: 'Requested Season does not exist. Cannot create Arc.'
             };
         }
 
         // Get the number of seasons available for this comic to set the season number
-        var count = await session.query(Arc).count({seasonId: body.seasonId}).asPromise();
+        const count = await session.query(Arc).count({seasonId: body.seasonId}).asPromise();
 
         // Create a new season
-        var arc = new Arc();
+        const arc = new Arc();
         arc.seasonId = body.seasonId;
         arc.title = body.title;
         arc.number = count + 1;
 
         // Save the season
-        await new Promise((resolve, reject) =>
-        {
+        await new Promise((resolve, reject) => {
             session.save(arc, () => resolve());
             session.flush();
         });
@@ -54,19 +50,17 @@ class Worker implements IRestWorker<Arc>
         };
     }
 
-    public async read(id: string): Promise<IPayload<Arc>>
-    {
-        var session = SessionManager.createSession();
-        var arc = await session.query(Arc)
+    public async read(id: string): Promise<IPayload<Arc>> {
+        const session = SessionManager.createSession();
+        let arc = await session.query(Arc)
             .findOne({_id: id})
             .asPromise();
 
-        if (!arc)
-        {
+        if (!arc) {
             return {
                 success: false,
                 message: `Arc with ID '${id}' does not exist.`
-            }
+            };
         }
 
         session.close();
@@ -74,26 +68,23 @@ class Worker implements IRestWorker<Arc>
         return {
             success: true,
             data: arc
-        }
+        };
     }
 
-    public async update(id: string, body): Promise<IPayload<Arc>>
-    {
+    public async update(id: string, body): Promise<IPayload<Arc>> {
         return undefined;
     }
 
-    public async remove(id: string): Promise<IPayload<Arc>>
-    {
+    public async remove(id: string): Promise<IPayload<Arc>> {
         return undefined;
     }
 
-    public async getArcsForSeason(seasonId: string, page = 1, show = 10): Promise<IPayload<Arc[]>>
-    {
-        var session = SessionManager.createSession();
+    public async getArcsForSeason(seasonId: string, page = 1, show = 10): Promise<IPayload<Arc[]>> {
+        const session = SessionManager.createSession();
 
-        var arcs = await session.query(Arc)
+        const arcs = await session.query(Arc)
             .findAll({seasonId})
-            .sort("number", 1)
+            .sort('number', 1)
             .skip(show * (page - 1))
             .limit(show)
             .asPromise();
@@ -103,7 +94,7 @@ class Worker implements IRestWorker<Arc>
         return {
             success: true,
             data: arcs
-        }
+        };
     }
 }
 
