@@ -1,47 +1,43 @@
-import {IPayload} from "../../../interfaces/IPayload";
-import {IRestWorker} from "../../../interfaces/IRestWorker";
-import {SessionManager} from "../../../database/SessionManager";
-import {Arc} from "../../../database/models/comic/Arc";
-import {Chapter} from "../../../database/models/comic/Chapter";
+import {IPayload} from '../../../interfaces/IPayload';
+import {IRestWorker} from '../../../interfaces/IRestWorker';
+import {SessionManager} from '../../../database/SessionManager';
+import {Arc} from '../../../database/models/comic/Arc';
+import {Chapter} from '../../../database/models/comic/Chapter';
 
-class Worker implements IRestWorker<Chapter>
-{
-    public async create(body: Chapter): Promise<IPayload<Chapter>>
-    {
+class Worker implements IRestWorker<Chapter> {
+
+    public async create(body: Chapter): Promise<IPayload<Chapter>> {
         // Validate request
-        if (!body.arcId)
-        {
+        if (!body.arcId) {
             return {
                 success: false,
-                message: "Arc information given is invalid. Cannot create Chapter."
+                message: 'Arc information given is invalid. Cannot create Chapter.'
             };
         }
 
-        var session = SessionManager.createSession();
+        const session = SessionManager.createSession();
 
         // Get the proper parent
-        var arc = await session.query(Arc).findOne({_id: body.arcId}).asPromise();
+        let arc = await session.query(Arc).findOne({_id: body.arcId}).asPromise();
 
-        if (!arc)
-        {
+        if (!arc) {
             return {
                 success: false,
-                message: "Requested Arc does not exist. Cannot create Chapter."
+                message: 'Requested Arc does not exist. Cannot create Chapter.'
             };
         }
 
         // Get the number of seasons available for this comic to set the season number
-        var count = await session.query(Chapter).count({arcId: body.arcId}).asPromise();
+        const count = await session.query(Chapter).count({arcId: body.arcId}).asPromise();
 
         // Create a new season
-        var chapter = new Chapter();
+        const chapter = new Chapter();
         chapter.arcId = body.arcId;
         chapter.title = body.title;
         chapter.number = count + 1;
 
         // Save the season
-        await new Promise((resolve, reject) =>
-        {
+        await new Promise((resolve, reject) => {
             session.save(chapter, () => resolve());
             session.flush();
         });
@@ -54,17 +50,15 @@ class Worker implements IRestWorker<Chapter>
         };
     }
 
-    public async read(id: string): Promise<IPayload<Chapter>>
-    {
-        var session = SessionManager.createSession();
-        var chapter = await session.query(Chapter).findOne({_id: id}).asPromise();
+    public async read(id: string): Promise<IPayload<Chapter>> {
+        const session = SessionManager.createSession();
+        let chapter = await session.query(Chapter).findOne({_id: id}).asPromise();
 
-        if (!chapter)
-        {
+        if (!chapter) {
             return {
                 success: false,
                 message: `Chapter with ID '${id}' does not exist.`
-            }
+            };
         }
 
         session.close();
@@ -72,26 +66,23 @@ class Worker implements IRestWorker<Chapter>
         return {
             success: true,
             data: chapter
-        }
+        };
     }
 
-    public async update(id: string, body): Promise<IPayload<Chapter>>
-    {
+    public async update(id: string, body): Promise<IPayload<Chapter>> {
         return undefined;
     }
 
-    public async remove(id: string): Promise<IPayload<Chapter>>
-    {
+    public async remove(id: string): Promise<IPayload<Chapter>> {
         return undefined;
     }
 
-    public async getChaptersForArc(arcId: string, page = 1, show = 10): Promise<IPayload<Chapter[]>>
-    {
-        var session = SessionManager.createSession();
+    public async getChaptersForArc(arcId: string, page = 1, show = 10): Promise<IPayload<Chapter[]>> {
+        const session = SessionManager.createSession();
 
-        var chapters = await session.query(Chapter)
+        const chapters = await session.query(Chapter)
             .findAll({arcId})
-            .sort("number", 1)
+            .sort('number', 1)
             .skip(show * (page - 1))
             .limit(show)
             .asPromise();
@@ -101,7 +92,7 @@ class Worker implements IRestWorker<Chapter>
         return {
             success: true,
             data: chapters
-        }
+        };
     }
 }
 
