@@ -1,19 +1,17 @@
-import {IPayload} from "../../../interfaces/IPayload";
-import {IRestWorker} from "../../../interfaces/IRestWorker";
-import {SessionManager} from "../../../database/SessionManager";
-import {Comic} from "../../../database/models/comic/Comic";
-import {Season} from "../../../database/models/comic/Season";
+import {IPayload} from '../../../interfaces/IPayload';
+import {IRestWorker} from '../../../interfaces/IRestWorker';
+import {SessionManager} from '../../../database/SessionManager';
+import {Comic} from '../../../database/models/comic/Comic';
+import {Season} from '../../../database/models/comic/Season';
 
-class Worker implements IRestWorker<Season>
-{
-    public async create(body: Season): Promise<IPayload<Season>>
-    {
+class Worker implements IRestWorker<Season> {
+
+    public async create(body: Season): Promise<IPayload<Season>> {
         // Validate request
-        if (!body.comicId)
-        {
+        if (!body.comicId) {
             return {
                 success: false,
-                message: "Comic information given is invalid. Cannot create Season."
+                message: 'Comic information given is invalid. Cannot create Season.'
             };
         }
 
@@ -22,11 +20,10 @@ class Worker implements IRestWorker<Season>
         // Get the proper comic
         var comic = await session.query(Comic).findOne({_id: body.comicId}).asPromise();
 
-        if (!comic)
-        {
+        if (!comic) {
             return {
                 success: false,
-                message: "Requested Comic does not exist. Cannot create Season."
+                message: 'Requested Comic does not exist. Cannot create Season.'
             };
         }
 
@@ -40,8 +37,7 @@ class Worker implements IRestWorker<Season>
         season.number = count + 1;
 
         // Save the season
-        await new Promise((resolve, reject) =>
-        {
+        await new Promise((resolve, reject) => {
             session.save(season, () => resolve());
             session.flush();
         });
@@ -54,17 +50,15 @@ class Worker implements IRestWorker<Season>
         };
     }
 
-    public async read(id: string): Promise<IPayload<Season>>
-    {
+    public async read(id: string): Promise<IPayload<Season>> {
         var session = SessionManager.createSession();
         var season = await session.query(Season).findOne({_id: id}).asPromise();
 
-        if (!season)
-        {
+        if (!season) {
             return {
                 success: false,
                 message: `Season with ID '${id}' does not exist.`
-            }
+            };
         }
 
         session.close();
@@ -72,26 +66,23 @@ class Worker implements IRestWorker<Season>
         return {
             success: true,
             data: season
-        }
+        };
     }
 
-    public async update(id: string, body): Promise<IPayload<Season>>
-    {
+    public async update(id: string, body): Promise<IPayload<Season>> {
         return undefined;
     }
 
-    public async remove(id: string): Promise<IPayload<Season>>
-    {
+    public async remove(id: string): Promise<IPayload<Season>> {
         return undefined;
     }
 
-    public async getSeasonsForComic(comicId: string, page = 1, show = 10): Promise<IPayload<Season[]>>
-    {
+    public async getSeasonsForComic(comicId: string, page = 1, show = 10): Promise<IPayload<Season[]>> {
         var session = SessionManager.createSession();
 
         var seasons = await session.query(Season)
             .findAll({comicId})
-            .sort("title", 1)
+            .sort('title', 1)
             .skip(show * (page - 1))
             .limit(show)
             .asPromise();
@@ -101,7 +92,7 @@ class Worker implements IRestWorker<Season>
         return {
             success: true,
             data: seasons
-        }
+        };
     }
 }
 
