@@ -1,5 +1,4 @@
-export class Model
-{
+export class Model {
     public _id: string;
 
     public id: string;
