@@ -1,5 +1,5 @@
 export const Roles = {
-    Administrator: "Administrator",
-    Editor: "Editor",
-    User: "User"
+    Administrator: 'Administrator',
+    Editor: 'Editor',
+    User: 'User'
 };
