@@ -1,7 +1,7 @@
-export * from "./auth/User";
-export * from "./auth/Passport";
-export * from "./comic/Comic";
-export * from "./comic/Season";
-export * from "./comic/Arc";
-export * from "./comic/Chapter";
-export * from "./comic/Page";
+export * from './auth/User';
+export * from './auth/Passport';
+export * from './comic/Comic';
+export * from './comic/Season';
+export * from './comic/Arc';
+export * from './comic/Chapter';
+export * from './comic/Page';
