@@ -1,10 +1,9 @@
-import {Field, Entity, ElementType} from "hydrate-mongodb";
-import {Model} from "../../base/Model";
-import {Chapter} from "./Chapter";
+import {Field, Entity, ElementType} from 'hydrate-mongodb';
+import {Model} from '../../base/Model';
+import {Chapter} from './Chapter';
 
 @Entity()
-export class Arc extends Model
-{
+export class Arc extends Model {
     @Field()
     public seasonId: string;
 
