@@ -1,7 +1,5 @@
-import {Season} from "../models/comic/Season";
-import {SessionManager} from "../SessionManager";
-import {Comic} from "../models/comic/Comic";
+import {SessionManager} from '../SessionManager';
+import {Comic} from '../models/comic/Comic';
 
-export class ComicRepository
-{
+export class ComicRepository {
 }
