@@ -1,7 +1,6 @@
-export interface IPayload<TData>
-{
-    success?:boolean;
-    data?:TData;
-    message?:string;
-    error?:string;
+export interface IPayload<TData> {
+    success?: boolean;
+    data?: TData;
+    message?: string;
+    error?: string;
 }
