@@ -1,8 +1,7 @@
-import {IPayload} from "./IPayload";
-import {Model} from "../database/base/Model";
+import {IPayload} from './IPayload';
+import {Model} from '../database/base/Model';
 
-export interface IRestWorker<TModel extends Model>
-{
+export interface IRestWorker<TModel extends Model> {
     create(body: any, ...args): Promise<IPayload<TModel>>;
 
     read(id: string, ...args): Promise<IPayload<TModel>>;
