@@ -1,14 +1,12 @@
-import * as passport from "passport";
-import * as session from "express-session";
+import * as passport from 'passport';
+import * as session from 'express-session';
 
-import {Express} from "express";
-import {LOCAL_STRATEGY} from "./PassportLocalConfig";
-import {CONFIG} from "../../config/Config";
+import {Express} from 'express';
+import {LOCAL_STRATEGY} from './PassportLocalConfig';
+import {CONFIG} from '../../config/Config';
 
-export class PassportConfig
-{
-    public static init(app: Express)
-    {
+export class PassportConfig {
+    public static init(app: Express) {
         app.use(session({
             secret: CONFIG.jwt.secret,
             resave: true,
@@ -17,12 +15,10 @@ export class PassportConfig
         app.use(passport.initialize());
         app.use(passport.session());
 
-        passport.serializeUser((user, callback) =>
-        {
+        passport.serializeUser((user, callback) => {
             callback(null, user);
         });
-        passport.deserializeUser((user, callback) =>
-        {
+        passport.deserializeUser((user, callback) => {
             callback(null, user);
         });
 
