@@ -1,19 +1,17 @@
-import {Strategy} from "passport-local";
-import {AuthWorker} from "../../api/workers/AuthWorker";
+import {Strategy} from 'passport-local';
+import {AuthWorker} from '../../api/workers/AuthWorker';
 
-async function strategyHandler(identifier, password, next)
-{
+async function strategyHandler(identifier, password, next) {
     // Let's check if the user input was valid
-    var loginResult = await AuthWorker.validateLogin(identifier, password);
-    if (loginResult.error)
-    {
+    const loginResult = await AuthWorker.validateLogin(identifier, password);
+    if (loginResult.error) {
         return next(null, null, {message: loginResult.error});
     }
 
-    var passportResult = loginResult.data;
+    const passportResult = loginResult.data;
 
     // Login the user and give them a session
     return next(null, passportResult);
 }
 
-export const LOCAL_STRATEGY = new Strategy({usernameField: "email", passwordField: "password"}, strategyHandler);
+export const LOCAL_STRATEGY = new Strategy({usernameField: 'email', passwordField: 'password'}, strategyHandler);
